+++ state_space.cpp
@@ -4,31 +4,31 @@
 #include <iostream>
 
 using namespace std;
-StateSpace c2d(const StateSpace& ss, float dt)
+StateSpace c2d(const StateSpace& ss, double dt)
 {
-    MatrixXf A = ss.A;
-    MatrixXf B = ss.B;
-    MatrixXf C = ss.C;
-    MatrixXf D = ss.D;
+    MatrixXd A = ss.A;
+    MatrixXd B = ss.B;
+    MatrixXd C = ss.C;
+    MatrixXd D = ss.D;
 
     // Compute matrix exponential
-    MatrixXf M(A.rows()+ B.cols(), A.cols() + B.cols()); // Augmented matrix for finding ZOH discretization
+    MatrixXd M(A.rows()+ B.cols(), A.cols() + B.cols()); // Augmented matrix for finding ZOH discretization
 
     M << A, B,
-         MatrixXf::Zero(B.cols(), A.cols()), MatrixXf::Zero(B.cols(), B.cols());
+         MatrixXd::Zero(B.cols(), A.cols()), MatrixXd::Zero(B.cols(), B.cols());
 
-    MatrixXf M_exp = (M*dt).exp();
+    MatrixXd M_exp = (M*dt).exp();
     // cout << "M exp matrix:\n" << M_exp << endl;
 
     // cout << "M exp matrix:\n" << M_exp << endl;
 
 
     // Discretize the continuous state space model
-    // MatrixXf Ad = (A*dt).exp();
-    MatrixXf Ad = M_exp.topLeftCorner(A.rows(), A.cols());
-    MatrixXf Bd = M_exp.topRightCorner(B.rows(), B.cols());
-    MatrixXf Cd = C;
-    MatrixXf Dd = D;
+    // MatrixXd Ad = (A*dt).exp();
+    MatrixXd Ad = M_exp.topLeftCorner(A.rows(), A.cols());
+    MatrixXd Bd = M_exp.topRightCorner(B.rows(), B.cols());
+    MatrixXd Cd = C;
+    MatrixXd Dd = D;
 
     return StateSpace(Ad, Bd, Cd, Dd);
 
@@ -36,10 +36,10 @@ StateSpace c2d(const StateSpace& ss, float dt)
 
 // int main(int, char**)
 // {   
-//     Eigen::MatrixXf A(2, 2);
-//     Eigen::MatrixXf B(2, 1);
-//     Eigen::MatrixXf C(1, 2);
-//     Eigen::MatrixXf D(1, 1);
+//     Eigen::MatrixXd A(2, 2);
+//     Eigen::MatrixXd B(2, 1);
+//     Eigen::MatrixXd C(1, 2);
+//     Eigen::MatrixXd D(1, 1);
 
 //     A << 0, 1,
 //          -2, -3;
@@ -50,7 +50,7 @@ StateSpace c2d(const StateSpace& ss, float dt)
 
 //     StateSpace go2_ss(A, B, C, D);
 
-//     float dt = 0.01;
+//     double dt = 0.01;
 //     StateSpace go2_ss_d = c2d(go2_ss, dt);
 
 //     std::cout << "Ad matrix:\n" << ss_discrete.A << std::endl;
